STARTED TREE-VIEW, things are looking good so far
