--- app/render/css/general.styl
@@ -1,34 +1,0 @@
-*
-  padding 0px
-  margin 0px
-  font-family sans-serif
-  outline none
-
-body
-  background base3
-
-container()
-  display flex
-
-container-col()
-  container()
-  flex-direction column
-
-container-row()
-  container()
-  flex-direction row
-
-.content
-  container-col()
-  align-content stretch
-  // display flex
-  // justify-content center
-  // align-items center
-
-.content-row
-  container-row()
-
-.wrapper
-  position absolute
-  width 100%
-  height 100%
