+++ app/render/css/layout.styl
@@ -1,0 +1,91 @@
+*
+  padding 0
+  margin 0
+  outline 0
+
+html
+body
+  font-family sans-serif
+  background base3
+  height 100%
+
+prattle-content
+  display block
+  height 100%
+  overflow scroll
+  position relative
+  background base3
+
+prattle-layout-axis
+  display flex
+
+prattle-pane
+  position relative
+  display flex
+  flex 1
+  padding 1em
+
+.horizontal
+  height 100%
+  flex 1
+  flex-flow row
+
+.vertical
+  height 100%
+  flex 1
+  flex-flow column
+
+.right
+  justify-content flex-start
+
+.left
+  justify-content flex-end
+
+.top
+  justify-content flex-start
+
+.bottom
+  justify-content flex-end;
+
+.center
+  justify-content center
+
+.full
+  height 100%
+  width 100%
+
+
+
+
+
+
+
+
+
+
+
+// container()
+//   display flex
+//
+// container-col()
+//   container()
+//   flex-direction column
+//
+// container-row()
+//   container()
+//   flex-direction row
+//
+// .content
+//   container-col()
+//   align-content stretch
+//   // display flex
+//   // justify-content center
+//   // align-items center
+//
+// .content-row
+//   container-row()
+//
+// .wrapper
+//   position absolute
+//   width 100%
+//   height 100%
