+++ app/render/css/style.css
@@ -1,1 +1,1 @@
-*{padding:0;margin:0;font-family:sans-serif;outline:none}body{background:#fdf6e3}.content{display:flex;flex-direction:column;align-content:stretch}.content-row{display:flex;flex-direction:row}.wrapper{position:absolute;width:100%;height:100%}.logo-container{display:flex;flex-direction:column;background:#fdf6e3;color:#002b36}.login-form{display:flex;flex-direction:column;align-items:center}.login-input{width:50%;font-size:20px;font-color:#002b36;border:none;border-bottom:#002b36 1px solid;padding:.5em 1.5em;background:#93a1a1}.login-logo-container{width:200px;height:120px}.tab-container{position:fixed;top:0;right:0;left:0;display:flex;flex-direction:row;background:#586e75;height:30px;border-bottom:4px solid #93a1a1}.tab{background:#93a1a1;border:none;border-left:#839496 3px solid;border-bottom:#839496 1px solid;display:flex;flex-direction:row;align-content:center;padding:10px}.tab::highlight{display:none}.tab-title{color:#073642;text-align:center}.tab-title::before{content:attr(data-name);font-size:initial}.tab-close-icon{padding-left:5px;border:none;background:none;color:#586e75;font-size:14px;height:100%}.tab-close-icon:hover{color:#dc322f}.tab-close-icon::after{content:'x'}.tab-active-true{color:#586e75;background:#fdf6e3;border-bottom:none;border-left:#b58900 3px solid}.tab-join{border:none;background:#93a1a1;font-size:20px;padding-left:5px;padding-right:5px;color:#002b36;border:#586e75 solid 2px;border-bottom:none;margin-left:5px}.tab-join::highlight{display:none}#messages-container{overflow:scroll;padding-top:30px;padding-bottom:75px;overflow:scroll}#input-container{border-top:2px solid #586e75;position:fixed;bottom:0;left:0;right:0;padding:15px;background:#839496}#messages{font-size:20px;list-style-type:none;}#messages li:nth-child(odd){background:#eee8d5}.message{color:3}#input-form{display:flex;flex-direction:row;justify-content:space-around;width:100%}#input-text{flex-grow:10;width:80%;margin:3px;font-size:20px;border:2px solid #586e75;background:#eee8d5}#input-button{flex-grow:1;margin:3px;font-size:20px;border:2px solid #002b36;border-radius:7px;background:#073642;color:#839496;padding:5px}#input-button::highlight{display:none}.annoucment{color:#586e75;display:table;margin:0 auto 0 auto}.nick{font-family:monospace;padding-left:20px}.nick::before{content:'<'}.nick::after{content:'>'}.self{color:#268bd2}.other{color:#2aa198}
+*{padding:0;margin:0;outline:0}html,body{font-family:sans-serif;background:#fdf6e3;height:100%}prattle-content{display:block;height:100%;overflow:scroll;position:relative;background:#fdf6e3}prattle-layout-axis{display:flex}prattle-pane{position:relative;display:flex;flex:1;padding:1em}.horizontal{height:100%;flex:1;flex-flow:row}.vertical{height:100%;flex:1;flex-flow:column}.right{justify-content:flex-start}.left{justify-content:flex-end}.top{justify-content:flex-start}.bottom{justify-content:flex-end}.center{justify-content:center}.full{height:100%;width:100%}.logo-container{background:#fdf6e3;color:#002b36}.login-form{align-items:center}.login-input{width:50%;font-size:20px;font-color:#002b36;border:none;border-bottom:#002b36 1px solid;padding:.5em 1.5em;background:#93a1a1}.login-logo-container{width:200px;height:120px}.tab-container{position:fixed;top:0;right:0;left:0;background:#586e75;height:30px;border-bottom:4px solid #93a1a1}.tab{background:#93a1a1;border:none;border-left:#839496 3px solid;border-bottom:#839496 1px solid;display:flex;flex-direction:row;align-content:center;padding:10px}.tab::highlight{display:none}.tab-title{color:#073642;text-align:center}.tab-title::before{content:attr(data-name);font-size:initial}.tab-close-icon{padding-left:5px;border:none;background:none;color:#586e75;font-size:14px;height:100%}.tab-close-icon:hover{color:#dc322f}.tab-close-icon::after{content:'x'}.tab-active-true{color:#586e75;background:#fdf6e3;border-bottom:none;border-left:#b58900 3px solid}.tab-join{border:none;background:#93a1a1;font-size:20px;padding-left:5px;padding-right:5px;color:#002b36;border:#586e75 solid 2px;border-bottom:none;margin-left:5px}.tab-join::highlight{display:none}.messages-view{display:flex;flex-grow:5;overflow:scroll}.messages-container{flex-grow:10;overflow:scroll}.input-container{flex-grow:1}prattle-message{color:#002b36;list-style-type:none;font-size:1em}.message{color:#002b36}.message-text,.nick{font-family:sans-serif}.nick::before{content:'<'}.nick::after{content:'>'}.self-true{color:#268bd2}.self-false{color:#2aa198}prattle-annoucment{color:#586e75;display:table;margin:0 auto}.tree-view{display:flex;flex-grow:1;background:#839496;padding-left:.5em;padding-right:.5em}prattle-channel{color:#073642}
