+++ app/render/css/style.styl
@@ -1,6 +1,7 @@
 @import './variables'
-@import './general'
+@import './layout'
 @import './login'
 @import './tabs'
 @import './messages'
 @import './join'
+@import './tree-view'
