+++ app/render/css/tree-view.styl
@@ -1,0 +1,9 @@
+.tree-view
+  display flex
+  flex-grow 1
+  background base0
+  padding-left 0.5em
+  padding-right 0.5em
+
+prattle-channel
+  color base02
