--- app/render/js/chat/directives/message.js
@@ -1,9 +1,0 @@
-chat.directive('message', function() {
-  return {
-    restrict: 'E',
-    scope: {
-
-    },
-    templateUrl: `file://${__dirname}/../../../templates/chat/tab.html`
-  };
-});
