--- app/render/js/chat/directives/tab.js
@@ -1,9 +1,0 @@
-chat.directive('tab', function() {
-  return {
-    restrict: 'E',
-    scope: {
-
-    },
-    templateUrl: `file://${__dirname}/../../../templates/chat/tab.html`
-  };
-});
