+++ app/render/js/controllers/chat-controller.js
@@ -1,4 +1,3 @@
-var chat = angular.module('chat', []);
 chat.controller('ChatController', function ($scope) {
   'use strict';
   var $ = require('../lib/jquery-1.11.3.min.js');
@@ -105,88 +104,3 @@ chat.controller('ChatController', function ($scope) {
   $scope.testMessage();
   $scope.testMessage();
 })
-
-.controller('ChannelController', function ($scope) {
-  'use strict';
-  var $ = require('../lib/jquery-1.11.3.min.js');
-  var ipc = require('ipc');
-  var remote = require('remote');
-  var mainWindow = remote.getCurrentWindow();
-  var client = mainWindow.client;
-
-  $scope.currentMessage = '';
-
-  // $scope.testMessage();
-
-  /////////////////////////
-  //BIND TO CLIENT EVENTS//
-  /////////////////////////
-
-  ipc.on('client-error', function (error) {
-    $scope.message($scope.active, 'error: ', error);
-    console.log(error);
-  });
-
-  ipc.on('client-motd', function (motd) {
-    $scope.announce(motd);
-  });
-
-  ipc.on('client-selfMessage', function (to, text) {
-    $scope.message(to, client.nick, text);
-  });
-
-  ipc.on('client-message', function (nick, to, text, message) {
-    $scope.message(to, nick, text);
-  });
-
-  ipc.on('client-names', function (names) {
-
-  });
-
-  ipc.on('client-kill', function (nick, reason, channels, message) {
-    // $scope.announce();
-  });
-
-  ipc.on('client-pm', function (nick, text, message) {
-
-  });
-
-  ipc.on('client-kick', function () {
-    // $scope.announce('You were kicked.. :(');
-  });
-
-  ipc.on('', function () {
-
-  });
-})
-
-.directive('chatPane', function () {
-  return {
-    template: '<div class="content wrapper">\n'+
-      '<div id="messages-container">\n'+
-        '<ul id="messages">\n'+
-          '<li class="message" ng-repeat="message in chat.messages">\n'+
-            '<span class="nick {{message.typeNick}}">{{message.nick}}</span>\n'+
-            '<span class="{{message.typeMessage}}">{{message.text}}</span>\n'+
-          '</li>\n'+
-        '</ul>\n'+
-      '</div>\n'+
-      '<div id="input-container">\n'+
-        '<form id="input-form" ng-submit="chat.submitMessage()" >\n'+
-          '<input type="text" id="input-text" ng-model="chat.currentMessage">\n'+
-          '<button type="button" id="input-button" ng-click="chat.submitMessage()">send</button>\n'+
-        '</form>\n'+
-      '</div>\n'+
-    '</div>'
-  };
-})
-
-.directive('chatTab', function () { //not in use currently
-  return {
-    template:
-    '<button ng-repeat="channel in chat.channels" class="tab tab-active-{{channel.active}}">\n'+
-      '<div class="tab-title" data-name="{{channel.name}}"></div>\n'+
-      '<div class="tab-close-icon"></div>\n'+
-    '</button>'
-  };
-});
