+++ app/render/pages/login.html
@@ -10,7 +10,6 @@
   <div class="content" ng-controller="LoginController as login">
     <div class="logo-container">
       <img class="login-logo-img" src="../assets/img/logo.jpg" alt="logo isn't loading 😖" />
-      <!-- <h1>Logo</h1> -->
     </div>
     <div class="login-container">
       <form class="login-form" id="login-form">
@@ -28,7 +27,7 @@
 
         <button ng-click="login.connectionTry()" class="login-input">connect</button>
         <!-- replace this  ^^^^^^^^^^^^^^^^^^^ to login.connectionTry() when done testing -->
-        <button ng-click="login.connectionReady()" class="login-input">don't connect</button>
+        <!-- <button ng-click="login.connectionReady()" class="login-input">don't connect</button> -->
       </form>
     </div>
   </div>
