+++ app/render/templates/chat/message.html
@@ -1,0 +1,7 @@
+<li class="message">
+  <!-- Uncomment for debugging -->
+  <!-- nick="{{nick}}" self="{{self}}" type="{{type}}" text="{{text}}"
+  <br> -->
+  <span class="nick self-{{self}}">{{nick}}</span>
+  <span class="message-text">{{text}}</span>
+</li>
