+++ app/render/templates/chat/tab.html
@@ -1,0 +1,6 @@
+<!-- <div ng-repeat="(name, channel) in channels" class="tab tab-active-{{channel.active}}"
+  ng-click="makeActive(name)">
+  <div class="tab-title" data-name="{{name}}"></div>
+  <button type="button" ng-click="leaveChannel(name)" class="tab-close-icon"></button>
+</div> -->
+<h1>I AM A TAB</h1>
